feat(render): one line per component, spinning while Figma works
A render printed the full spec path before starting and a second line
after finishing, so a batch scrolled past as two lines per component with
the same long path prefix on every other one.

It now spins in place — `Rendering: dsButton` with a live elapsed count —
and the outcome overwrites it: `✓ Rendered: dsButton (3.5s, nodeId: …)`.
The shape matches what fetch already sets for per-item work, and drops
"in Figma", which no other command says of itself.

The spinner moves out of FetchCommand into a shared utility rather than
being copied. Outside a TTY it prints its text once and still reports
elapsed time, so logs and CI stay free of escape codes, and it stops in a
`finally` so a failed render cannot leave it spinning.

Writer warnings are withheld for now behind a single constant. They are
dominated by known, tracked defects and by degradations a user cannot act
on, so a wall of them per render buries the outcome. Failures and the
count of unresolved instance elements still print — a broken render cannot
pass as a clean one.
